+++ src/views/MeetingDetail/TalkDetailContainer2024.tsx
@@ -5,12 +5,13 @@ import SectionWrapper from "../../components/SectionWrapper/SectionWrapper";
 import styled from "styled-components";
 import {useParams} from "react-router";
 import conferenceData from "../../data/2024.json";
-import {sessionAdapter, useFetchTalksById} from "../Talks/UseFetchTalks";
+import {useFetchTalksById} from "../Talks/UseFetchTalks";
 import * as Sentry from "@sentry/react";
 import {useFetchSpeakers} from "../Speakers/UseFetchSpeakers";
 import {Session} from "../Talks/Talk.types";
 import MeetingDetail from "./MeetingDetail";
 import {ISpeaker} from "../../types/speakers";
+import {sessionAdapter} from "../../services/sessionsAdapter";
 
 const StyledContainer = styled.div`
   background-color: ${Color.WHITE};
